requirements.txt
Django==2.1.7
django-tastypie==0.14.2
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2018.9
six==1.12.0
